moved App.kt and UIUtil.kt
