+++ app/src/main/java/org/guru/playlistmaker/app/App.kt
@@ -1,4 +1,4 @@
-package org.guru.playlistmaker.util
+package org.guru.playlistmaker.app
 
 import android.app.Application
 import org.guru.playlistmaker.creator.Creator
